Dial back type checking + remove concurrency module
